Create binance_RSI Example.py
Playing with some RSI's I built this little buyer and seller of ETH which will place a buy order and maybe 2 if the conditions are met. Buy ETH with USDT. It also will place a sell order after successful buy @ a .5% bump. At this point the variables are inline and should be brought to the top of the document for easy editing. But it works well.